refactor: parseCarNames 메서드 view 입력 기능으로 이동
