+++ src/main/java/racingcar/domain/car/RacingCarName.java
@@ -14,16 +14,4 @@ public class RacingCarName {
         return this.name;
     }
 
-    public static List<String> parseCarNames(String carNamesInput) {
-        List<String> carNames = Arrays.asList(carNamesInput.split(","));
-        for (String carName : carNames) {
-            validateCarName(carName);
-        }
-        return carNames;
-    }
-
-    private static void validateCarName(String carName) {
-        if (carName.isEmpty() || carName.length() > 5)
-            throw new IllegalArgumentException("차 이름은 1~5 글자로 입력해주세요.");
-    }
 }
